Build-in libpng now uses same zlib headers as rest of Irrlicht We never modified libpng which just used #include "zlib.h" so far. The reason it worked mostly is that we added include paths to zlib all over the Makefile and project files. So it used the one in the zlib subfolder anyways - ignoring the _IRR_USE_NON_SYSTEM_ZLIB_ define. Except xcode project file never had that include. So now OSX has a different version than Irrlicht and that version difference mattered in libpng compiling broke.
Note: All the adding of zlib (and in Makefile even all other sub-libraries) to the include path would really not be necessary
with the current Irrlicht solution. So I could remove those now. All it does is mess up _IRR_USE_NON_SYSTEM_ZLIB_ even more.
But it won't break anything which wasn't already broken, so not changing it in this patch.
Going to comment it some more in another patch... and probably rewrite it for Irrlicht 1.9


git-svn-id: https://svn.code.sf.net/p/irrlicht/code/branches/releases/1.8@6512 dfc29bdd-3216-0410-991c-e03cc46cb475